requirements.txt
fastapi
uvicorn
pymongo
passlib[bcrypt]
python-dotenv
pyjwt
pandas
joblib
